MB-57888: Index Update (#64)
- Added UpdateFieldInfo which tracks which parts of the zapx fields have
been deleted during an update
 - Added a new index interface which implements two new functions
- OpenMeta opens only the index metadata allowing us to read index
definition
without starting the index
- UpdateFields allows us to write to meta all of the updated fields info
while
the index is closed

Any updates to the index needs to be done while the index is closed.
This is to ensure that once the index is live, the maps will not be
updated under any circumstances. Hot reloads will require mapping which
we are trying to avoid as these maps are used during query paths.